Use the configured culture when writing cell values
This makes `TestIssue312()` pass on a computer where the current culture is a French culture. Assert.Equal("12,345.68", rows[1].A) would fail because the value would be serialized as `12 345,68` instead of `12,345.68`.

Also simplify the implementation of the Excel and CSV writers by getting rid of both `ExcelFormatToStringMethod` and `ExcelcultureToStringMethod` which are replaced by testing if the value implements the `IFormattable` interface.